pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "ytb-summarizer"
version = "1.0.0"
description = "YouTube video summarizer with Qt GUI and multi-LLM support"
requires-python = ">=3.11"
dependencies = [
    "youtube-transcript-api>=1.2.0",
    "yt-dlp>=2024.1.1",
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "PySide6>=6.7.0",
    "markdown>=3.6",
    "pyyaml>=6.0",
    "platformdirs>=4.0",
]

[project.scripts]
ytb-summarizer = "main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["ytb_summarizer*", "ytb_gui*"]
